# api/resource/event/event.go
package event

import (
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"time"

	"github.com/go-playground/validator/v10"
	"github.com/google/uuid"
	"github.com/jmoiron/sqlx"
	"github.com/ushiradineth/cron-be/models"
	"github.com/ushiradineth/cron-be/util/event"
	"github.com/ushiradineth/cron-be/util/response"
	"github.com/ushiradineth/cron-be/util/user"
)

type API struct {
	db        *sqlx.DB
	validator *validator.Validate
}

func New(db *sqlx.DB, validator *validator.Validate) *API {
	return &API{
		db:        db,
		validator: validator,
	}
}

// @Summary		Get Event by ID
// @Description	Get authenticated user's event based on the JWT and event ID sent with the request
// @Tags			Event
// @Produce		json
// @Param			Path	path		EventPathParams	true	"EventPathParams"
// @Success		200		{object}	response.Response{data=models.Event}
// @Failure		400		{object}	response.Error
// @Failure		401		{object}	response.Error
// @Failure		500		{object}	response.Error
// @Security		BearerAuth
// @Router			/events/{event_id} [get]
func (api *API) Get(w http.ResponseWriter, r *http.Request) {
	path := EventPathParams{
		EventID: r.PathValue("event_id"),
	}

	if err := api.validator.Struct(path); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	user := user.GetUserFromJWT(r, w, api.db)
	if user == nil {
		return
	}

	event := event.GetEvent(w, path.EventID, user.ID.String(), api.db)
	if event == nil {
		return
	}

	log.Printf("Event %s has been retrieved by user %s", path.EventID, user.ID)

	response.HTTPResponse(w, event)
}

// @Summary		Create Event
// @Description	Create Event based on the parameters sent with the request
// @Tags			Event
// @Accept			json
// @Produce		json
// @Param			Body	body		EventBodyParams	true	"EventBodyParams"
// @Success		200		{object}	response.Response{data=models.Event}
// @Failure		400		{object}	response.Error
// @Failure		401		{object}	response.Error
// @Failure		500		{object}	response.Error
// @Security		BearerAuth
// @Router			/events [post]
func (api *API) Post(w http.ResponseWriter, r *http.Request) {
	var body EventBodyParams
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	if err := api.validator.Struct(body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	user := user.GetUserFromJWT(r, w, api.db)
	if user == nil {
		return
	}

	eventExists := event.DoesEventExist("", body.StartTime, body.EndTime, user.ID.String(), api.db)

	parsedStart, err := time.Parse(time.RFC3339, body.StartTime)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	parsedEnd, err := time.Parse(time.RFC3339, body.EndTime)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	if eventExists {
		response.HTTPError(w, http.StatusBadRequest, "Event already exists", response.StatusFail)
		return
	}

	if parsedStart.After(parsedEnd) {
		response.HTTPError(w, http.StatusBadRequest, "Start time must not be after end time", response.StatusFail)
		return
	}

	event := models.Event{
		ID:       uuid.New(),
		Title:    body.Title,
		Start:    parsedStart,
		End:      parsedEnd,
		UserID:   user.ID,
		Timezone: body.Timezone,
		Repeated: body.Repeated,
	}

	_, err = api.db.Exec("INSERT INTO events (id, title, start_time, end_time, user_id, timezone, repeated) VALUES ($1, $2, $3, $4, $5, $6, $7)", event.ID, event.Title, event.Start, event.End, event.UserID, event.Timezone, event.Repeated)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	log.Printf("Event %s has been created by user %s", event.ID, user.ID)

	response.HTTPResponse(w, event)
}

// @Summary		Update Event
// @Description	Update Event based on the parameters sent with the request
// @Tags			Event
// @Accept			json
// @Produce		json
// @Param			Path	path		EventPathParams	true	"EventPathParams"
// @Param			Body	body		EventBodyParams	true	"EventBodyParams"
// @Success		200		{object}	response.Response{data=models.Event}
// @Failure		400		{object}	response.Error
// @Failure		401		{object}	response.Error
// @Failure		500		{object}	response.Error
// @Security		BearerAuth
// @Router			/events/{event_id} [put]
func (api *API) Put(w http.ResponseWriter, r *http.Request) {
	path := EventPathParams{
		EventID: r.PathValue("event_id"),
	}

	if err := api.validator.Struct(path); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	var body EventBodyParams
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	if err := api.validator.Struct(body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	user := user.GetUserFromJWT(r, w, api.db)
	if user == nil {
		return
	}

	existingEvent := event.GetEvent(w, path.EventID, user.ID.String(), api.db)
	if existingEvent == nil {
		response.HTTPError(w, http.StatusBadRequest, "Event does not exists", response.StatusFail)
		return
	}

	parsedStart, err := time.Parse(time.RFC3339, body.StartTime)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	parsedEnd, err := time.Parse(time.RFC3339, body.EndTime)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	if parsedStart.After(parsedEnd) {
		response.HTTPError(w, http.StatusBadRequest, "Start time must not be after end time", response.StatusFail)
		return
	}

	parsedUUID, err := uuid.Parse(path.EventID)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	event := models.Event{
		ID:       parsedUUID,
		Title:    body.Title,
		Start:    parsedStart,
		End:      parsedEnd,
		UserID:   user.ID,
		Timezone: body.Timezone,
		Repeated: body.Repeated,
	}

	_, err = api.db.Exec("UPDATE events SET title=$1, start_time=$2, end_time=$3, timezone=$4, repeated=$5 WHERE id=$6 AND user_id=$7", event.Title, event.Start, event.End, event.Timezone, event.Repeated, event.ID, event.UserID.String())
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	log.Printf("Event %s has been updated by user %s", path.EventID, user.ID)

	response.HTTPResponse(w, event)
}

// @Summary		Delete Event
// @Description	Delete Event based on the parameters sent with the request
// @Tags			Event
// @Produce		json
// @Param			Path	path		EventPathParams	true	"EventPathParams"
// @Success		200		{object}	response.Response{data=string}
// @Failure		400		{object}	response.Error
// @Failure		401		{object}	response.Error
// @Failure		500		{object}	response.Error
// @Security		BearerAuth
// @Router			/events/{event_id} [delete]
func (api *API) Delete(w http.ResponseWriter, r *http.Request) {
	path := EventPathParams{
		EventID: r.PathValue("event_id"),
	}

	if err := api.validator.Struct(path); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	user := user.GetUserFromJWT(r, w, api.db)
	if user == nil {
		return
	}

	res, err := api.db.Exec("DELETE FROM events WHERE id=$1 AND user_id=$2", path.EventID, user.ID.String())
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	count, err := res.RowsAffected()
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	if count == 0 {
		response.GenericBadRequestError(w, fmt.Errorf("Event does not exist"))
		return
	}

	log.Printf("Event %s has been deleted by user %s", path.EventID, user.ID)

	response.HTTPResponse(w, "Event has been successfully deleted")
}

// @Summary		Get User Events
// @Description	Get authenticated user's event based on the JWT sent with the request
// @Tags			Event
// @Accept			json
// @Produce		json
// @Param			Body	body		GetUserEventsBodyParams	true	"GetUserEventsBodyParams"
// @Success		200		{object}	response.Response{data=[]models.Event}
// @Failure		400		{object}	response.Error
// @Failure		401		{object}	response.Error
// @Failure		500		{object}	response.Error
// @Security		BearerAuth
// @Router			/events [get]
func (api *API) GetUserEvents(w http.ResponseWriter, r *http.Request) {
	var body GetUserEventsBodyParams
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	if err := api.validator.Struct(body); err != nil {
		response.GenericValidationError(w, err)
		return
	}

	user := user.GetUserFromJWT(r, w, api.db)
	if user == nil {
		return
	}

	parsedStart, err := time.Parse("2006-01-02", body.StartDay)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	parsedEnd, err := time.Parse("2006-01-02", body.EndDay)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	events := []models.Event{}

	err = api.db.Select(&events, "SELECT * FROM events WHERE user_id=$1 AND start_time >= $2 AND start_time <= $3", user.ID, parsedStart, parsedEnd)
	if err != nil {
		response.GenericServerError(w, err)
		return
	}

	log.Printf("Events for user %s have been retrieved", user.ID)

	response.HTTPResponse(w, events)
}